+++ emborg/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '1.4.1'
-__released__ = '2019-04-24'
+__version__ = '1.4.2'
+__released__ = '2019-09-22'
 
 from .settings import Settings as Emborg